Extract construction of cached axios from create infra
