+++ src/infrastructure/create-infrastructure.ts
@@ -1,5 +1,3 @@
-import axios from 'axios';
-import { AxiosCacheInstance, buildStorage, setupCache } from 'axios-cache-interceptor';
 import { Json } from 'fp-ts/Json';
 import * as O from 'fp-ts/Option';
 import * as RA from 'fp-ts/ReadonlyArray';
@@ -7,9 +5,9 @@ import * as T from 'fp-ts/Task';
 import * as TE from 'fp-ts/TaskEither';
 import { identity, pipe } from 'fp-ts/function';
 import { Pool } from 'pg';
-import { createClient } from 'redis';
 import { Adapters } from './adapters';
 import { commitEvents, writeEventToDatabase } from './commit-events';
+import { createCachedAxios } from './create-cached-axios';
 import { fetchDataset } from './fetch-dataset';
 import { fetchHypothesisAnnotation } from './fetch-hypothesis-annotation';
 import { fetchNcrcReview } from './fetch-ncrc-review';
@@ -134,10 +132,11 @@ export const createInfrastructure = (dependencies: Dependencies): TE.TaskEither<
       logger: lowLevelAdapters.logger,
     }),
   })),
+  TE.apS('cachedAxios', TE.fromTask(createCachedAxios)),
   TE.chain((partialAdapters) => TE.tryCatch(
     async () => {
       const {
-        events, logger, pool, getJson,
+        events, logger, pool, getJson, cachedAxios,
       } = partialAdapters;
 
       const getAllEvents = T.of(events);
@@ -168,36 +167,6 @@ export const createInfrastructure = (dependencies: Dependencies): TE.TaskEither<
         addArticleToList({ getAllEvents, commitEvents: commitEventsWithoutListeners, ...partialAdapters }),
         TE.map(() => undefined),
       );
-
-      let cachedAxios: AxiosCacheInstance;
-
-      if (process.env.APP_CACHE === 'redis') {
-        const client = createClient({
-          socket: { host: 'sciety_cache' },
-        });
-
-        await client.connect();
-
-        const redisStorage = buildStorage({
-          async find(key) {
-            const result = await client.get(`axios-cache:${key}`);
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-            return (result !== null ? JSON.parse(result) : undefined);
-          },
-
-          async set(key, value) {
-            await client.set(`axios-cache:${key}`, JSON.stringify(value));
-          },
-
-          async remove(key) {
-            await client.del(`axios-cache:${key}`);
-          },
-        });
-
-        cachedAxios = setupCache(axios, { storage: redisStorage });
-      } else {
-        cachedAxios = setupCache(axios);
-      }
 
       return {
         fetchArticle: fetchCrossrefArticle(
